+++ workalendar/tests/test_registry.py
@@ -1,34 +1,20 @@
-from datetime import date
 from unittest import TestCase
 
 from workalendar.registry import IsoRegistry
 from workalendar.core import Calendar
+from workalendar.exceptions import ISORegistryError
 
 
 class RegionCalendar(Calendar):
     'Region'
 
-    def holidays(self, year=None):
-        return tuple((
-            (date(year, 12, 25), 'Christmas'),
-            (date(year, 1, 1), 'New year'),
-        ))
-
-    def get_weekend_days(self):
-        return []  # no week-end, yes, it's sad
-
 
 class SubRegionCalendar(Calendar):
     'Sub Region'
 
-    def holidays(self, year=None):
-        return tuple((
-            (date(year, 12, 25), 'Christmas'),
-            (date(year, 1, 1), 'New year'),
-        ))
 
-    def get_weekend_days(self):
-        return []  # no week-end, yes, it's sad
+class NotACalendarClass(object):
+    "Not a Calendar"
 
 
 class MockCalendarTest(TestCase):
@@ -71,3 +57,8 @@ class MockCalendarTest(TestCase):
         items = registry.items(['RE'], include_subregions=False)
         self.assertEqual(1, len(items))
         self.assertIn('RE', items)
+
+    def test_register_non_calendar(self):
+        registry = IsoRegistry(load_standard_modules=False)
+        with self.assertRaises(ISORegistryError):
+            registry.register("NAC", NotACalendarClass)